# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BookinApi",
        "description": "Booking Api description",
        "contact": {
            "email": "bmihaylov99@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/rooms/{id}": {
            "get": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Get a specific room",
                "description": "Get a specific room.",
                "operationId": "4be01e7c39a22026b3129b34b0795bf3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the room",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "number": {
                                                    "type": "string",
                                                    "example": "101"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "Standard"
                                                },
                                                "price_per_night": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 100.5
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "available"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Room not found"
                    }
                }
            }
        },
        "/api/rooms": {
            "post": {
                "tags": [
                    "Rooms"
                ],
                "summary": "Store a new room",
                "description": "Store a new room.",
                "operationId": "cd6c1029f22e2425a1f99cf83e27383a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "number",
                                    "type",
                                    "price_per_night",
                                    "status"
                                ],
                                "properties": {
                                    "number": {
                                        "type": "string",
                                        "example": "101"
                                    },
                                    "type": {
                                        "type": "string",
                                        "example": "Standard"
                                    },
                                    "price_per_night": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 100.5
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "available"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Room created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "number": {
                                                    "type": "string",
                                                    "example": "101"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "Standard"
                                                },
                                                "price_per_night": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 100.5
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "available"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Room created successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Failed to authenticate"
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "number": [
                                                    "The number field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        }
    },
...
}